backend/requirements.txt
django==3.1.4
psycopg2
djangorestframework==3.12.2